move overmap.cpp inline structs to omdata.h
