+++ src/mapgen_functions.cpp
@@ -21,6 +21,7 @@
 #include "cuboid_rectangle.h"
 #include "flood_fill.h"
 #include "map.h"
+#include "mapdata.h"
 #include "map_iterator.h"
 #include "map_scale_constants.h"
 #include "mapgen.h"
